feat: agregar verificación de permisos en rutas y componentes, mejora…

+++ routes/web.php
@@ -38,7 +38,7 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::get('dashboard', function () {
         return Inertia::render('dashboard');
     })->name('dashboard');
-     Route::get('usuarios', [UserController::class, 'Index'])->name('usuarios.Index');
+     Route::get('usuarios', [UserController::class, 'Index'])->name('usuarios.Index')->middleware('permission:view usuarios|edit usuarios|create usuarios');
      Route::patch('/usuarios/{user}/toggle-status', [UserController::class, 'toggleStatus'])
     ->name('usuarios.toggle-status');
     Route::get('usuarios/create', [UserController::class, 'create'])->name('usuarios.Create');
@@ -47,7 +47,7 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::put('usuarios/{user}',[UserController::class,'update'])->name('usuarios.Update');
     Route::delete('usuarios/{user}',[UserController::class,'destroy'])->name('usuarios.Destroy');
     //rutas de rol y permisos
-    Route::get('rol', [RoleController::class, 'index'])->name('rol.Index');
+    Route::get('rol', [RoleController::class, 'index'])->name('rol.Index')->middleware('permission:view roles|edit roles|create roles');
     Route::get('rol/edit/{role}', [RoleController::class, 'edit'])->name('rol.Editar');
     Route::put('rol/{role}', [RoleController::class, 'update'])->name('rol.Update');
     Route::get('rol/create', [RoleController::class, 'create'])->name('rol.Create');
@@ -56,12 +56,12 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::post('rol/bulk-delete', [RoleController::class, 'bulkDelete'])->name('rol.BulkDestroy');
 
     //rutas de permisos
-     Route::get('permisos', [PermisoController::class, 'index'])->name('permisos.Index');
-    Route::get('permisos/create', [PermisoController::class, 'create'])->name('permisos.Create');
-    Route::post('permisos', [PermisoController::class, 'store'])->name('permisos.Store');
-    Route::get('permisos/edit/{permission}', [PermisoController::class, 'edit'])->name('permisos.Editar');
-    Route::put('permisos/{permission}', [PermisoController::class, 'update'])->name('permisos.Update');
-    Route::delete('permisos/{permission}', [PermisoController::class, 'destroy'])->name('permisos.Destroy');
+     Route::get('permisos', [PermisoController::class, 'index'])->name('permisos.Index')->middleware('permission:view permisos|edit permisos|create permisos');
+    Route::get('permisos/create', [PermisoController::class, 'create'])->name('permisos.Create')->middleware('permission:create permisos');
+    Route::post('permisos', [PermisoController::class, 'store'])->name('permisos.Store')->middleware('permission:create permisos');
+    Route::get('permisos/edit/{permission}', [PermisoController::class, 'edit'])->name('permisos.Editar')->middleware('permission:edit permisos');
+    Route::put('permisos/{permission}', [PermisoController::class, 'update'])->name('permisos.Update')->middleware('permission:edit permisos');
+    Route::delete('permisos/{permission}', [PermisoController::class, 'destroy'])->name('permisos.Destroy')->middleware('permission:delete permisos');
 
     //rutas de aula
     Route::get('aula', [AulaController::class, 'index'])->name('aula.Index')->middleware('permission:view aulas|edit aulas|create aulas');
@@ -73,28 +73,28 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::delete('aula/{classroom}', [AulaController::class, 'destroy'])->name('aula.Destroy');
 
         //rutas de materias
-    Route::get('/materias', [MateriaController::class, 'index'])->name('materias.index');
-    Route::get('/materias/create', [MateriaController::class, 'create'])->name('materias.create');
-    Route::post('/materias', [MateriaController::class, 'store'])->name('materias.store');
-    Route::get('/materias/{materia}/edit', [MateriaController::class, 'edit'])->name('materias.edit');
-    Route::put('/materias/{materia}', [MateriaController::class, 'update'])->name('materias.update');
-    Route::patch('/materias/{materia}/change-status', [MateriaController::class, 'changeStatus'])->name('materias.change-status');
+    Route::get('/materias', [MateriaController::class, 'index'])->name('materias.index')->middleware('permission:view materias|edit materias|create materias');
+    Route::get('/materias/create', [MateriaController::class, 'create'])->name('materias.create')->middleware('permission:create materias');
+    Route::post('/materias', [MateriaController::class, 'store'])->name('materias.store')->middleware('permission:create materias');
+    Route::get('/materias/{materia}/edit', [MateriaController::class, 'edit'])->name('materias.edit')->middleware('permission:edit materias');
+    Route::put('/materias/{materia}', [MateriaController::class, 'update'])->name('materias.update')->middleware('permission:edit materias');
+    Route::patch('/materias/{materia}/change-status', [MateriaController::class, 'changeStatus'])->name('materias.change-status')->middleware('permission:edit materias');
 
     // Rutas de grupos
-    Route::get('/grupos', [GrupoController::class, 'index'])->name('grupos.index');
-    Route::get('/grupos/create', [GrupoController::class, 'create'])->name('grupos.create');
-    Route::post('/grupos', [GrupoController::class, 'store'])->name('grupos.store');
-    Route::get('/grupos/{grupo}/edit', [GrupoController::class, 'edit'])->name('grupos.edit');
-    Route::put('/grupos/{grupo}', [GrupoController::class, 'update'])->name('grupos.update');
-    Route::patch('/grupos/{grupo}/change-status', [GrupoController::class, 'changeStatus'])->name('grupos.change-status');
+    Route::get('/grupos', [GrupoController::class, 'index'])->name('grupos.index')->middleware('permission:view grupos|edit grupos|create grupos');
+    Route::get('/grupos/create', [GrupoController::class, 'create'])->name('grupos.create')->middleware('permission:create grupos');
+    Route::post('/grupos', [GrupoController::class, 'store'])->name('grupos.store')->middleware('permission:create grupos');
+    Route::get('/grupos/{grupo}/edit', [GrupoController::class, 'edit'])->name('grupos.edit')->middleware('permission:edit grupos');
+    Route::put('/grupos/{grupo}', [GrupoController::class, 'update'])->name('grupos.update')->middleware('permission:edit grupos');
+    Route::patch('/grupos/{grupo}/change-status', [GrupoController::class, 'changeStatus'])->name('grupos.change-status')->middleware('permission:edit grupos');
 
     // Rutas de docentes
- Route::get('/docentes', [DocenteController::class, 'index'])->name('docentes.index');
-    Route::get('/docentes/create', [DocenteController::class, 'create'])->name('docentes.create');
-    Route::post('/docentes', [DocenteController::class, 'store'])->name('docentes.store');
-    Route::get('/docentes/{docente}/edit', [DocenteController::class, 'edit'])->name('docentes.edit');
-    Route::put('/docentes/{docente}', [DocenteController::class, 'update'])->name('docentes.update');
-    Route::delete('/docentes/{docente}', [DocenteController::class, 'destroy'])->name('docentes.destroy');
+ Route::get('/docentes', [DocenteController::class, 'index'])->name('docentes.index')->middleware('permission:view docentes|edit docentes|create docentes');
+    Route::get('/docentes/create', [DocenteController::class, 'create'])->name('docentes.create')->middleware('permission:create docentes');
+    Route::post('/docentes', [DocenteController::class, 'store'])->name('docentes.store')->middleware('permission:create docentes');
+    Route::get('/docentes/{docente}/edit', [DocenteController::class, 'edit'])->name('docentes.edit')->middleware('permission:edit docentes');
+    Route::put('/docentes/{docente}', [DocenteController::class, 'update'])->name('docentes.update')->middleware('permission:edit docentes');
+    Route::delete('/docentes/{docente}', [DocenteController::class, 'destroy'])->name('docentes.destroy')->middleware('permission:delete docentes');
     Route::patch('/docentes/{docente}/status', [DocenteController::class, 'changeStatus'])->name('docentes.changeStatus');
     
     // Rutas adicionales
@@ -103,12 +103,12 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::get('/docentes/{docente}', [DocenteController::class, 'show'])->name('docentes.show');
 
         // // Rutas adicionales
-    Route::get('/inventario/movimientos', [MovimientoInventarioController::class, 'index'])->name('inventario.movimientos.index');
+    Route::get('/inventario/movimientos', [MovimientoInventarioController::class, 'index'])->name('inventario.movimientos.index')->middleware('permission:view movimientos');
     Route::get('/inventario/movimientos/registrar', [MovimientoInventarioController::class, 'create'])->name('inventario.movimientos.create');
     Route::post('/inventario/movimientos', [MovimientoInventarioController::class, 'store'])->name('inventario.movimientos.store');
 
     // Rutas de inventario
-    Route::get('/inventario', [InventarioController::class, 'index'])->name('inventario.index');
+    Route::get('/inventario', [InventarioController::class, 'index'])->name('inventario.index')->middleware('permission:view inventario|edit inventario|create inventario');
     Route::get('/inventario/create', [InventarioController::class, 'create'])->name('inventario.create');
     Route::post('/inventario', [InventarioController::class, 'store'])->name('inventario.store');
     Route::get('/inventario/{inventario}/edit', [InventarioController::class, 'edit'])->name('inventario.edit');
@@ -122,7 +122,7 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::get('/inventario/{inventario}/detalle', [DetalleInventarioController::class, 'show'])->name('inventario.detalle.show');
 
     // Rutas Periodo Académico
-    Route::get('/periodos-academicos', [PeriodoAcademicoController::class, 'index'])->name('periodos-academicos.index');
+    Route::get('/periodos-academicos', [PeriodoAcademicoController::class, 'index'])->name('periodos-academicos.index')->middleware('permission:view periodos|edit periodos|create periodos');
     Route::get('/periodos-academicos/create', [PeriodoAcademicoController::class, 'create'])->name('periodos-academicos.create');
     Route::post('/periodos-academicos', [PeriodoAcademicoController::class, 'store'])->name('periodos-academicos.store');
     Route::get('/periodos-academicos/{periodosAcademico}/edit', [PeriodoAcademicoController::class, 'edit'])->name('periodos-academicos.edit');
@@ -134,12 +134,12 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::get('/periodos-academicos/{periodosAcademico}/grupos', [PeriodoAcademicoController::class, 'grupos'])->name('periodos-academicos.grupos');
     Route::get('/periodos-academicos/{periodosAcademico}/detalle', [PeriodoAcademicoController::class, 'detalle'])->name('periodos-academicos.detalle');
     // //rutas de horario docente
-    Route::get('/horario', [HorarioDocenteController::class, 'index'])->name('horario.index');
+    Route::get('/horario', [HorarioDocenteController::class, 'index'])->name('horario.index')->middleware('permission:view docentes|edit docentes|create docentes');
     Route::get('/horario/semanal', [HorarioDocenteController::class, 'horarioSemanal'])->name('horario.semanal');
     Route::get('/horario/materia/{materia}', [HorarioDocenteController::class, 'showMateria'])->name('horario.materia');
 
 
-    Route::get('/asistencia', [AsistenciaController::class, 'index'])->name('asistencia.index');
+    Route::get('/asistencia', [AsistenciaController::class, 'index'])->name('asistencia.index')->middleware('permission:view docentes|edit docentes|create docentes');
     Route::get('/historial', [AsistenciaController::class, 'historial'])->name('asistencia.historial');
     Route::post('/asistencia/generar-qr/{horario}', [AsistenciaController::class, 'generarQR'])->name('asistencia.generar-qr');
     Route::post('/asistencia/justificar/{horario}', [AsistenciaController::class, 'justificarFalta'])->name('asistencia.justificar');
@@ -152,7 +152,7 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::get('/bitacora/exportar', [BitacoraController::class, 'exportar'])->name('bitacora.exportar');
 
     // Rutas horarios
-    Route::get('bloques-horarios', [BloqueHorarioController::class, 'index'])->name('bloques-horarios.index');
+    Route::get('bloques-horarios', [BloqueHorarioController::class, 'index'])->name('bloques-horarios.index')->middleware('permission:view bloques|edit bloques|create bloques');
     Route::get('bloques-horarios/create', [BloqueHorarioController::class, 'create'])->name('bloques-horarios.create');
     Route::post('bloques-horarios', [BloqueHorarioController::class, 'store'])->name('bloques-horarios.store');
     Route::get('bloques-horarios/{bloquesHorario}', [BloqueHorarioController::class, 'show'])->name('bloques-horarios.show');
@@ -166,7 +166,7 @@ Route::middleware(['auth', 'verified', 'user.active', 'track.activity'])->group(
     Route::get('bloques-horarios/{bloquesHorario}/aulas', [BloqueHorarioController::class, 'aulas'])->name('bloques-horarios.aulas');
 
     // Rutas de Asignaciones
-    Route::get('/asignaciones', [AsignacionController::class, 'index'])->name('asignaciones.index');
+    Route::get('/asignaciones', [AsignacionController::class, 'index'])->name('asignaciones.index')->middleware('permission:view asignaciones');
     Route::get('/asignaciones/create', [AsignacionController::class, 'create'])->name('asignaciones.create');
     Route::post('/asignaciones', [AsignacionController::class, 'store'])->name('asignaciones.store');
     Route::get('/asignaciones/{asignacion}/edit', [AsignacionController::class, 'edit'])->name('asignaciones.edit');
